+++ apps/client/src/components/product-card.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-import Tag from "$components/tag.svelte";
+import Tag from "$components/tag/tag.svelte";
 import Card from "$components/ui/card/card.svelte";
 import { formatDistanceToNow } from "date-fns";
 import PencilIcon from "lucide-svelte/icons/pencil";
