+++ apps/client/src/components/tag/index.ts
@@ -1,0 +1,3 @@
+import Tag from "./tag.svelte";
+
+export { Tag };
